Initial linking functionality
Links can be created by pressing <Ctrl-l>
Currently links are forgotten upon changing tab, but the next few
commits will be working on keeping them saved upon tab change and saving
them with the file.

There are two options for saving them with the note:
    Separate file - causes cluttered notes directory
    Appended to bottom of note

Appended to the bottom of the note looks to be the better solution, as
it is cleaner, but if the user inspects the note with a text editor they
may change the links.